cmd: Add a CLI helpers package
Add the cmd/helpers, helpers package. This a centralized location for CLI helpers that can be called in existing CLI tooling.

Update the cmd/[deploy-metering/reporting-operator] CLI tooling, move some of the helper functions stored in those main packages, into the CLI helpers package.
